+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2022-04-19
+
+### Deprecated
+- The `difference` alias for `symmetric_difference` has been deprecated.
+
+    To avoid conflicting with the *std* library where `difference` is equivalent
+    to *imbl*'s `relative_complement`.
+
+
 ## [2.0.0] - 2022-04-12
 
 ### Fixed